requirements.txt
pandas
pillow
fastapi>=0.111
uvicorn[standard]>=0.30
pydantic>=2.7